+++ src/main/java/com/global/api/entities/enums/PayByLinkType.java
@@ -1,5 +1,7 @@
 package com.global.api.entities.enums;
 
 public enum PayByLinkType {
-    PAYMENT
+    PAYMENT,
+    HOSTED_PAYMENT_PAGE,
+    THIRD_PARTY_PAGE
 }